+++ layer.py
@@ -20,7 +20,7 @@ class AgentLayer:
         """
         self.agents = agents
         self.nagents = len(agents)
-        self.layer_state = np.full((xs, ys), -20)
+        self.layer_state = np.full((xs, ys), -20.0, dtype=np.float16)
         # print("AgentLayer shape: ", self.layer_state.shape)
 
     def n_agents(self):
@@ -38,14 +38,14 @@ class AgentLayer:
         """Clear the old position and set the new position in the layer state."""
         ox, oy = tuple(map(int, old_position))
         nx, ny = tuple(map(int, new_position))
-        if self.layer_state[ox, oy] == 0:  # Only reset if it was the current position of the agent
+        if self.layer_state[ox, oy] == 0.0:  # Only reset if it was the current position of the agent
             self.layer_state[ox, oy] = -0.1  # Start decay from -0.1
-        self.layer_state[nx, ny] = 0  # Refresh the new position to 0
+        self.layer_state[nx, ny] = 0.0  # Refresh the new position to 0
 
     def set_position(self, agent_idx, x, y):
         o_pos = self.agents[agent_idx].current_position()
         self.agents[agent_idx].set_position(x, y)
-        n_pos = np.zeros(2, dtype=np.int16)
+        n_pos = np.zeros(2, dtype=np.float16)
         n_pos[0], n_pos[1] = x, y
         self.update_position(o_pos, n_pos)
 
@@ -62,7 +62,7 @@ class AgentLayer:
             pos = self.agents[agent_idx].current_position()
             self.agents.pop(agent_idx)
             self.nagents -= 1
-            self.layer_state[int(pos[0]), int(pos[1])] = -20  # Clear the position in the layer state
+            self.layer_state[int(pos[0]), int(pos[1])] = -20.0 # Clear the position in the layer state
 
     def get_state_matrix(self):
         """Returns a matrix representing the positions of all allies."""
@@ -70,21 +70,21 @@ class AgentLayer:
     
     def update(self):
         # Decay previous positions
-        mask = (self.layer_state <= 0) & (self.layer_state > -20)
+        mask = (self.layer_state <= 0.0) & (self.layer_state > -20.0)
         self.layer_state[mask] -= 0.1  # Decrement the state of previously occupied positions
         # Reset positions that were more than 20 time steps ago
-        self.layer_state[self.layer_state < -20] = -20
+        self.layer_state[self.layer_state < -20.0] = -20.0
         # Update positions based on current agent locations
         for agent in self.agents:
             x, y = tuple(map(int, agent.current_position()))
-            self.layer_state[int(x), int(y)] = 0  # Set current agent positions to 0
+            self.layer_state[x, y] = 0.0  # Set current agent positions to 0
     
     
 class TargetLayer:
     def __init__(self, xs, ys, targets, map_matrix, seed=None):
         self.targets = targets
         self.map_matrix = map_matrix
-        self.layer_state = np.full((xs, ys), -20)
+        self.layer_state = np.full((xs, ys), -20.0, dtype=np.float16)
         self.ntargets = len(targets)
 
     def get_position(self, target_idx):
@@ -102,24 +102,24 @@ class TargetLayer:
     def update_position(self, old_position, new_position):
         ox, oy = tuple(map(int, old_position))
         nx, ny = tuple(map(int, new_position))
-        if self.layer_state[ox, oy] == 0:  # Only reset if it was the current position of the target
+        if self.layer_state[ox, oy] == 0.0:  # Only reset if it was the current position of the target
             self.layer_state[ox, oy] = -0.1  # Start decay from -0.1
-        self.layer_state[nx, ny] = 0  # Set new position to 0 (target present)
+        self.layer_state[nx, ny] = 0.0  # Set new position to 0 (target present)
 
     def n_targets(self):
         return self.ntargets
     
     def update(self):
         # Decay previous positions
-        mask = (self.layer_state <= 0) & (self.layer_state > -20)
+        mask = (self.layer_state <= 0.0) & (self.layer_state > -20.0)
         self.layer_state[mask] -= 0.1  # Decrement the state of previously occupied positions
         # Reset positions that were more than 20 time steps ago
-        self.layer_state[self.layer_state < -20] = -20
+        self.layer_state[self.layer_state < -20.0] = -20.0
         # Update positions based on current agent locations
         for target in self.targets:
             #x, y = agent.current_position()
             x, y = tuple(map(int, target.current_position()))
-            self.layer_state[int(x), int(y)] = 0  # Set current agent positions to 0
+            self.layer_state[x, y] = 0.0  # Set current agent positions to 0
 
     def get_state_matrix(self):
         """Returns a matrix representing the positions of all allies."""
@@ -131,7 +131,7 @@ class JammerLayer:
     def __init__(self, xs, ys, jammers, activation_times=None, seed=1):
         self.jammers = jammers
         self.nagents = len(jammers)
-        self.layer_state = np.full((xs, ys), -20)
+        self.layer_state = np.full((xs, ys), -20.0, dtype=np.float16)
         self.activation_times = activation_times or [0] * len(jammers)  # Default to immediate activation
 
     def activate_jammers(self, current_time):
@@ -156,11 +156,11 @@ class JammerLayer:
         Updates the layer state matrix with current positions of all jammers. 
         0 in the matrix is a current position of an active jammer.
         """
-        self.layer_state.fill(-20)
+        self.layer_state.fill(-20.0)
         for jammer in self.jammers:
             x, y = jammer.current_position()
             if jammer.is_active():
-                self.layer_state[x, y] = 0
+                self.layer_state[x, y] = 0.0
 
     def get_state_matrix(self):
         """Returns a matrix representing the positions of active jammers."""
